ユーザー管理機能のリファクタリング（途中）
ユーザー登録機能のエラーハンドリング検討（途中）
