+++ BasketClub/src/main/java/jp/wat/basket/controller/UserEditController.java
@@ -64,157 +64,6 @@ public class UserEditController {
 	private PasswordEncoder passwordEncoder;
 	
 	/////////////////////////////////////////////////////////
-	// ユーザー登録
-	/////////////////////////////////////////////////////////
-	/**
-	 *  登録画面　表示
-	 * @param userInfo
-	 * @param model
-	 * @return
-	 */
-	@RequestMapping(value="/user/regist/input", method=RequestMethod.GET)
-	public String registInput(UserInfo userInfo, Model model){
-				
-		userInfo.setStartViewName("/user/regist/input");
-		
-		// セレクトボックスのItemを取得（チーム区分、学年、ロール）
-		model.addAttribute("selectTeam", EnumTeamKubun.values());
-		model.addAttribute("selectGrade", EnumGrade.values());
-		model.addAttribute("selectRole", EnumRole.values());
-		
-		// ログインユーザー情報取得
-		LoginUser loginUser = commonService.getLoginUser();
-		
-		model.addAttribute("userName", loginUser.getUserName());
-		model.addAttribute("userForm", new UserForm());
-		
-		return "/user/regist/userRegistInput";
-	}
-	
-	/**
-	 *  登録画面　キャンセル
-	 * @param model
-	 * @return
-	 */
-	@RequestMapping(value = "/user/regist/confirm", method = RequestMethod.POST, params = "cancel")
-	public String registCancel(Model model) {
-		return "redirect:/user";
-	}
-	
-	/**
-	 *  登録確認画面　表示　
-	 * @param userForm
-	 * @param result
-	 * @param redirectAttributes
-	 * @param model
-	 * @return
-	 */
-	@RequestMapping(value = "/user/regist/confirm", method = RequestMethod.POST, params = "confirm")
-	public String registComfirm(@Validated UserForm userForm,
-			BindingResult result,
-			RedirectAttributes redirectAttributes,
-			Model model) {
-		
-		try {
-			
-			// ユーザーIDの重複チェック
-			if (userService.findById(userForm.getUserId()) != null)  {
-				result.rejectValue("userId", null , "登録済みのIDです。"); 
-			}
-			
-			// パスワード/パスワード確認の入力チェック
-			Map<String, List<String>> errorMap = userForm.passwordValidate(true);
-			List<String> passwordErrorMsgList = errorMap.get("password") != null? errorMap.get("password"): new ArrayList<String>();
-			List<String> rePasswordErrorMsgList = errorMap.get("rePassword")!= null? errorMap.get("rePassword"): new ArrayList<String>();
-			
-			// パスワードエラーがあればエラーメッセージを設定
-			for(String passwordErrorMsg : passwordErrorMsgList){
-				result.rejectValue("password", null , passwordErrorMsg); 
-			}
-			
-			// パスワード（確認用）にエラーがあればエラーメッセージを設定
-			for(String rePasswordErrorMsg : rePasswordErrorMsgList){
-				result.rejectValue("rePassword", null , rePasswordErrorMsg); 
-			}
-			
-			if(result.hasErrors()){
-				// ロールセレクトボックスのItemを設定
-				model.addAttribute("selectRole", EnumRole.values());
-				return "/user/regist/userRegistInput";
-			}
-					
-		} catch (Exception exception) {
-			logger.error("[@registComfirm]userForm.userId=" + userForm.getUserId() +" errorMessage:" + exception.getMessage());
-			
-			//入力画面に遷移
-			model.addAttribute("selectRole", EnumRole.values());
-			model.addAttribute("userForm", userForm);
-			model.addAttribute("errorMessage", "想定外のエラーが発生しました。<br>操作をやり直してください。");
-			return "/user/regist/userRegistInput";
-		}
-				
-		// ロールの設定
-		model.addAttribute("EnumRole", EnumRole.decode(userForm.getRoleCode()));
-		userForm.setRole(EnumRole.decode(userForm.getRoleCode()).getSName());
-		
-		// ログインユーザー情報取得
-		LoginUser loginUser = commonService.getLoginUser();
-		
-		model.addAttribute("userName", loginUser.getUserName());
-		model.addAttribute("userForm", userForm);
-				
-		return "/user/regist/userRegistConfirm";
-	
-	}
-	
-	/**
-	 *  登録処理
-	 * @param form
-	 * @param userInfo
-	 * @param result
-	 * @param sessionStatus
-	 * @param model
-	 * @param redirectAttributes
-	 * @return
-	 */
-	@RequestMapping(value={"/user/regist/transactfinish"}, method=RequestMethod.POST, params="submit")
-	public String registComplete(@Validated UserForm form, UserInfo userInfo,BindingResult result, SessionStatus sessionStatus, Model model,
-			RedirectAttributes redirectAttributes){
-
-		ModelMapper modelMapper = new ModelMapper();
-		LoginUser user = modelMapper.map(form, LoginUser.class);
-				
-		// パスワード暗号化
-		user.setPassword(passwordEncoder.encode(user.getPassword()));
-		
-		// 共通項目の設定
-		//LoginUser loginUser = commonService.getLoginUser();
-		user.setDeleteFlg(0);
-		user.setRegistUser(user.getUserId());
-		user.setRegistTime(new Timestamp(System.currentTimeMillis()));
-		user.setUpdateUser(user.getUserId());
-		user.setUpdateTime(new Timestamp(System.currentTimeMillis()));
-
-		try{
-			// DB更新処理 
-			userService.registUser(user);
-		
-		} catch(RuntimeException exception){
-			logger.error("[@registComplete]userForm.userId=" + form.getUserId() +" errorMessage:" + exception.getMessage());
-			
-			//入力画面に遷移
-			model.addAttribute("selectRole", EnumRole.values());
-			model.addAttribute("userForm", form);
-			model.addAttribute("errorMessage", "想定外のエラーが発生しました。<br>操作をやり直してください。");
-			return "/user/regist/userRegistInput";
-		}
-
-		redirectAttributes.addFlashAttribute("message","登録が完了しました");
-		return "redirect:/user/regist/input";
-		
-	}	
-	
-	/////////////////////////////////////////////////////////
 	// ユーザー変更
 	/////////////////////////////////////////////////////////
 	/**
