+++ BasketClub/src/main/resources/templates/user/userDetail.html
@@ -132,7 +132,6 @@ $(document).ready( function() {
 		flashMessage($errMessage, "error");	
 	}
 
-
 });
 
 </script>
